# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cookey Gateway API
  version: 2.0.0
  description: >
    Self-hosted personal API gateway. App-facing routes authenticate with a
    grant bearer token (ck_…) or PoP headers; admin routes with the owner
    session cookie or ADMIN_SECRET bearer; cron routes with CRON_SECRET.
servers:
  - url: "{gatewayUrl}"
    variables:
      gatewayUrl:
        default: https://your-gateway.example.com
components:
  securitySchemes:
    grantToken:
      type: http
      scheme: bearer
      description: Grant bearer token (ck_…)
    popHeaders:
      type: apiKey
      in: header
      name: x-sig
      description: PoP v1 — x-pop-v, x-app-id, x-ts, x-nonce, x-sig (see docs/POP_PROTOCOL.md)
    adminSecret:
      type: http
      scheme: bearer
      description: ADMIN_SECRET (or the admin session cookie)
    cronSecret:
      type: http
      scheme: bearer
      description: CRON_SECRET
  schemas:
...
    GrantDocument:
      type: object
      description: See docs/GRANT_SPEC.md
      required: [specVersion, app, runtime, auth, requests, duration]
      properties:
        specVersion: { type: string, enum: ["1"] }
        app:
          type: object
          required: [name]
          properties:
            name: { type: string }
            description: { type: string }
            homepage: { type: string, format: uri }
            iconUrl: { type: string, format: uri }
        runtime: { type: string, enum: [server, serverless, cli, browser] }
        auth: { type: string, enum: [bearer, pop] }
        publicKey: { type: [string, "null"] }
        requests:
          type: array
          minItems: 1
          items:
            type: object
            required: [resource, actions, reason]
            properties:
              resource: { type: string, pattern: "^[a-z]+:([a-z0-9-]+|\\*)$" }
              actions: { type: array, items: { type: string } }
              reason: { type: string, maxLength: 300 }
              constraints: { type: object }
        duration: { type: string }
        renewal:
          type: object
          properties: { period: { type: string } }
        budget:
          type: object
          properties:
            dailyRequests: { type: integer }
            monthlyRequests: { type: integer }
            dailyTokens: { type: integer }
            monthlyTokens: { type: integer }
        redirectUri: { type: string, format: uri }
...
paths:
...
  /api/connect/prepare:
    post:
      summary: Submit a grant document with a pairing code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [connectCode, grant]
              properties:
                connectCode: { type: string }
                grant: { $ref: "#/components/schemas/GrantDocument" }
      responses:
        "200":
          description: PENDING grant created
          content:
            application/json:
              schema:
                type: object
                properties:
                  grantId: { type: string }
                  sessionToken: { type: string, description: "= grantId (alias used by /api/connect/status polling)" }
                  approvalUrl: { type: string, format: uri }
                  expiresAt: { type: string, format: date-time }
        "400": { description: Invalid code or document }
...
  /api/connect/rotate:
    post:
      summary: Rotate PoP credential (signed with the current key)
      security: [{ popHeaders: [] }]
      responses:
        "200": { description: New credential active }
  /api/app/status:
    get:
      summary: App self-status (PoP)
      security: [{ popHeaders: [] }]
      responses:
        "200": { description: App info, permissions, grant state }
...